+++ sure-core/src/main/java/com/sure/tool/date/DateUtil.java
@@ -908,4 +908,24 @@ public class DateUtil {
 	}
 
 
+
+	/**
+	 * 判断两个日期是否同年。
+	 *
+	 * @param date1 日期一
+	 * @param date2 日期二
+	 * @return 是否同年
+	 */
+	public static boolean isSameYear(java.util.Date date1, java.util.Date date2) {
+		if (date1 == null || date2 == null) {
+			return false;
+		}
+		java.util.Calendar cal1 = java.util.Calendar.getInstance();
+		cal1.setTime(date1);
+		java.util.Calendar cal2 = java.util.Calendar.getInstance();
+		cal2.setTime(date2);
+		return cal1.get(java.util.Calendar.YEAR) == cal2.get(java.util.Calendar.YEAR);
+	}
+
+
 }